// frontend/src/types/permissions.ts
export type Permission =
  | "view_dashboard"
  | "view_products"
  | "create_product"
  | "edit_product"
  | "delete_product"
  | "view_stock_movements"
  | "create_stock_adjustment"
  | "view_invoices"
  | "create_invoice"
  | "edit_invoice"
  | "view_invoice_detail"
  | "print_invoice"
  | "confirm_invoice"
  | "cancel_invoice"
  | "record_payment"
  | "view_quotations"
  | "create_quotation"
  | "edit_quotation"
  | "view_quotation_detail"
  | "print_quotation"
  | "convert_quotation"
  | "mark_quotation_sent"
  | "delete_quotation"
  | "view_customers"
  | "create_customer"
  | "edit_customer"
  | "view_customer_detail"
  | "view_payments"
  | "view_accounts_payable"
  | "view_reports"
  | "view_users"
  | "view_user_detail"
  | "create_user"
  | "edit_user"
  | "view_categories"
  | "create_category"
  | "edit_category"
  | "view_suppliers"
  | "view_supplier_detail"
  | "create_supplier"
  | "edit_supplier"
  | "view_activity_logs"
  | "view_settings";